Added ability to allocate TFCameraViewController with default interface from Swift
Non default initializers are only callable from Swift if they have parameters. This changes adds a static method to allocate and initialize a TFCameraViewController with the default interface.

This change deprecates the initWithInterface method, as a static method is the preferred way to allocate and initialize a class with default values.

This change also removes the duplicate allocate call in initWithInterface.